requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
DAWG-Python==0.7.2
filelock==3.16.1
fsspec==2024.9.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.1
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240930
pycparser==2.22
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
srt==3.5.3
sympy==1.13.3
tiktoken==0.8.0
torch==2.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
vosk==0.3.45
websockets==13.1
